examples/robodk/requirements.txt
robodk
scipy
numpy
Pillow

